+++ pkg/deployments/index.ts
@@ -1,5 +1,21 @@
-import { Contract } from 'ethers';
+import { Contract } from '@ethersproject/contracts';
 import { Artifact } from 'hardhat/types';
+
+/**
+ * @dev Returns the task id and contract name for a canonical contract deployed on a specific network.
+ * Throws if the address doesn't match any known Balancer deployment.
+ * @param address Address of the contract to be fetched
+ * @param network Name of the network looking the deployment for (e.g. mainnet,  polygon, goerli, etc)
+ */
+export function lookupBalancerContractByAddress(address: string, network: string): { task: string; name: string } {
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  const networkAddresses = require(getBalancerContractAddresses(network));
+  const deploymentInfo = networkAddresses[address];
+  if (deploymentInfo === undefined) {
+    throw new Error(`Unable to connect ${address} to any Balancer deployment on ${network}`);
+  }
+  return deploymentInfo;
+}
 
 /**
  * @dev Creates an ethers Contract object for a canonical contract deployed on a specific network
@@ -20,12 +36,11 @@ export async function getBalancerContract(task: string, contract: string, networ
  */
 export async function getBalancerContractAt(task: string, contract: string, address: string): Promise<Contract> {
   const artifact = getBalancerContractArtifact(task, contract);
-  const { ethers } = await import('hardhat');
-  return ethers.getContractAt(artifact.abi, address);
+  return new Contract(address, artifact.abi);
 }
 
 /**
- * @dev Returns the artifact for a contract from a specific task
+ * @dev Returns the contract's artifact from a specific task
  * @param task ID of the task to look the ABI of the required contract
  * @param contract Name of the contract to looking the ABI of
  */
@@ -91,4 +106,12 @@ function getBalancerContractArtifactPath(task: string, contract: string): string
  */
 function getBalancerDeploymentPath(task: string, network: string): string {
   return `@balancer-labs/v2-deployments/dist/tasks/${task}/output/${network}.json`;
+}
+
+/**
+ * @dev Returns the path for the list of Balancer contract addresses on a network
+ * @param network Name of the network looking the deployment path for (e.g. mainnet, polygon, goerli, etc)
+ */
+function getBalancerContractAddresses(network: string): string {
+  return `@balancer-labs/v2-deployments/dist/addresses/${network}.json`;
 }