+++ pvt/benchmarks/misc.ts
@@ -9,13 +9,9 @@ import Vault from '@balancer-labs/v2-helpers/src/models/vault/Vault';
 import { StablePoolEncoder, toNormalizedWeights, WeightedPoolEncoder } from '@balancer-labs/balancer-js';
 import { MAX_UINT256, ZERO_ADDRESS, MAX_WEIGHTED_TOKENS } from '@balancer-labs/v2-helpers/src/constants';
 import { bn } from '@balancer-labs/v2-helpers/src/numbers';
-import { advanceTime, MONTH, DAY } from '@balancer-labs/v2-helpers/src/time';
+import { advanceTime, MONTH } from '@balancer-labs/v2-helpers/src/time';
 import { range } from 'lodash';
-import {
-  BasePoolRights,
-  ManagedPoolParams,
-  ManagedPoolRights,
-} from '@balancer-labs/v2-helpers/src/models/pools/weighted/types';
+import { ManagedPoolParams } from '@balancer-labs/v2-helpers/src/models/pools/weighted/types';
 import { poolConfigs } from './config';
 import { ProtocolFee } from '@balancer-labs/v2-helpers/src/models/vault/types';
 
@@ -81,12 +77,15 @@ export async function deployPool(vault: Vault, tokens: TokenList, poolName: Pool
 
     switch (poolName) {
       case 'ManagedPool': {
-        const newPoolParams: ManagedPoolParams = {
+        const userInputs = {
           name: name,
           symbol: symbol,
+          assetManagers: Array(tokens.length).fill(ZERO_ADDRESS),
+        };
+
+        const managedPoolSettings: ManagedPoolParams = {
           tokens: tokens.addresses,
           normalizedWeights: weights,
-          assetManagers: Array(tokens.length).fill(ZERO_ADDRESS),
           swapFeePercentage: swapFeePercentage,
           swapEnabledOnStart: true,
           mustAllowlistLPs: false,
@@ -94,22 +93,7 @@ export async function deployPool(vault: Vault, tokens: TokenList, poolName: Pool
           aumFeeId: ProtocolFee.AUM,
         };
 
-        const basePoolRights: BasePoolRights = {
-          canTransferOwnership: true,
-          canChangeSwapFee: true,
-          canUpdateMetadata: true,
-        };
-
-        const managedPoolRights: ManagedPoolRights = {
-          canChangeWeights: true,
-          canDisableSwaps: true,
-          canSetMustAllowlistLPs: true,
-          canSetCircuitBreakers: true,
-          canChangeTokens: true,
-          canChangeMgmtFees: true,
-        };
-
-        params = [newPoolParams, basePoolRights, managedPoolRights, DAY, creator.address];
+        params = [userInputs, managedPoolSettings, creator.address];
         break;
       }
       default: {
@@ -207,16 +191,15 @@ async function deployPoolFromFactory(
   if (poolName == 'ManagedPool') {
     const addRemoveTokenLib = await deploy('v2-pool-weighted/ManagedPoolAddRemoveTokenLib');
     const circuitBreakerLib = await deploy('v2-pool-weighted/CircuitBreakerLib');
-    const baseFactory = await deploy('v2-pool-weighted/ManagedPoolFactory', {
-      args: [vault.address, vault.getFeesProvider().address],
+    factory = await deploy('v2-pool-weighted/ManagedPoolFactory', {
+      args: [vault.address, vault.getFeesProvider().address, 'factoryVersion', 'poolVersion'],
       libraries: {
         CircuitBreakerLib: circuitBreakerLib.address,
         ManagedPoolAddRemoveTokenLib: addRemoveTokenLib.address,
       },
     });
-    factory = await deploy('v2-pool-weighted/ControlledManagedPoolFactory', { args: [baseFactory.address] });
   } else if (poolName == 'ComposableStablePool') {
-    factory = await deploy(`${fullName}Factory`, { args: [vault.address, vault.getFeesProvider().address] });
+    factory = await deploy(`${fullName}Factory`, { args: [vault.address, vault.getFeesProvider().address, '', ''] });
   } else {
     factory = await deploy(`${fullName}Factory`, { args: [vault.address, vault.getFeesProvider().address] });
   }
@@ -228,7 +211,7 @@ async function deployPoolFromFactory(
 
   if (poolName == 'ManagedPool') {
     receipt = await (await factory.connect(args.from).create(...args.parameters)).wait();
-    event = receipt.events?.find((e) => e.event == 'ManagedPoolCreated');
+    event = receipt.events?.find((e) => e.event == 'PoolCreated');
   } else {
     receipt = await (await factory.connect(args.from).create(name, symbol, ...args.parameters, ZERO_ADDRESS)).wait();
     event = receipt.events?.find((e) => e.event == 'PoolCreated');